Implemented a Key Generator Feature. While this isn't in requirements, this is realistically how a internal tool like this would work for a company like SecureLog.
Now if you were actually using this tool, it would be best to use something like OpenSSL for generation and distribution. Only using this platform for encryption and decryption. But for our testing purposes I have added in a key generator into the UI for my personal use.

Still debating removing for final response, but we will cross that bridge once we get there