Add EncryptForm component for data encryption functionality, super basic (#1)
* Add EncryptForm component for data encryption functionality, super basic

Tasks
1. Improve error messages and PEM format specification
2. improve UI choices
3. Add a exponential wait time to reduce number of requests a user can send

* Added improved error messages for incyption for common PEM errors such as lack of header and footer, as well as a not base64 encoded message within that

* Add DecryptForm component and integrate it into the App
updated some colors as well, still messing with frontend scheme

* Add LogsTable and Pagination components for displaying and navigating request logs

* Implemented a Key Generator Feature. While this isn't in requirements, this is realistically how a internal tool like this would work for a company like SecureLog.

Now if you were actually using this tool, it would be best to use something like OpenSSL for generation and distribution. Only using this platform for encryption and decryption. But for our testing purposes I have added in a key generator into the UI for my personal use.

Still debating removing for final response, but we will cross that bridge once we get there

* style: auto-format Python code [skip ci]

---------

Co-authored-by: lukebrevoort <11137273+lukebrevoort@users.noreply.github.com>